requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiounitest==1.5.0
altair==6.0.0
antiorm==1.2.1
attrs==26.1.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
db==0.1.1
db-sqlite3==0.0.1
fake-useragent==2.2.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
howlongtobeatpy==1.0.21
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
Levenshtein==0.27.3
MarkupSafe==3.0.3
multidict==6.7.1
narwhals==2.19.0
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pillow==12.2.0
plotly==6.6.0
propcache==0.4.1
protobuf==7.34.1
pyarrow==23.0.1
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-Levenshtein==0.27.3
RapidFuzz==3.14.5
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
streamlit==1.56.0
tenacity==9.1.4
thefuzz==0.22.1
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
watchdog==6.0.0
wrapt==2.1.2
yarl==1.23.0
